:hammer: refactor: Model 분리
