+++ Gitodo/Sources/Model/LoginData.swift
@@ -1,0 +1,12 @@
+//
+//  LoginData.swift
+//  Gitodo
+//
+//  Created by jiyeon on 5/7/24.
+//
+
+import Foundation
+
+struct Token: Codable {
+    let accessToken: String
+}
